+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.7.0-beta.1 (Unreleased)
-
+- Added support for restoring blob container.
 
 ## 12.6.0 (2020-04-06)
 - Fixed a bug that would prevent client initialization against Azurite in some containerized environments.